+++ app/Http/Controllers/Dash/OrderController.php
@@ -3,23 +3,41 @@
 namespace App\Http\Controllers\Dash;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\OrderResource;
 use App\Models\Order;
-use Illuminate\Http\Request;
-use Carbon\Carbon;
 
 class OrderController extends Controller
 {
     public function index()
     {
-        Carbon::setLocale('ar');
         $orders = Order::all();
-        return view('admin.orders.index', compact('orders'));
+        if (count($orders) > 0) {
+            return response()->json([
+                'success' => true,
+                'orders' => OrderResource::collection($orders)
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'msg' => 'there is no orders yet'
+            ], 404);
+        }
     }
 
     public function destroy($id)
     {
         $order = Order::find($id);
-        $order->delete();
-        return redirect(route('admin.orders'))->with('success', 'تم حذف الطلب بنجاح');
+        if ($order) {
+            $order->delete();
+            return response()->json([
+                'success' => true,
+                'msg' => 'order has been deleted successfully'
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'msg' => 'there is no such order'
+            ], 404);
+        }
     }
 }
